apps/server/ai/requirements.txt
onnxruntime==1.16.3
onnx==1.15.0
flask==3.0.0
requests==2.31.0
redis==5.0.1
Pillow==10.1.0
numpy==1.24.3
tokenizers==0.15.0